Sync launcher script and enable concurrent runs (#128)
* Sync launcher script and enable concurrent runs


* Update launch_actions_runner.sh to work locally


* Fix comment


---------

+++ .github/launch_actions_runner.sh
@@ -5,7 +5,6 @@
 #SBATCH --mem=32G
 #SBATCH --gpus=1
 #SBATCH --time=00:30:00
-#SBATCH --dependency=singleton
 #SBATCH --output=logs/runner_%j.out
 
 ## This script can be used to launch a new self-hosted GitHub runner.
@@ -17,16 +16,15 @@ set -o errexit
 # $SLURM_TMPDIR won't be set.
 set -o nounset
 
+# Seems to be required for the `uvx` to be found. (adds $HOME/.cargo/bin to PATH)
+source $HOME/.cargo/env
+# This is where the SH_TOKEN secret environment variable is set.
+source $HOME/.bash_aliases
 
 readonly repo="mila-iqia/ResearchTemplate"
 readonly action_runner_version="2.317.0"
 readonly expected_checksum_for_version="9e883d210df8c6028aff475475a457d380353f9d01877d51cc01a17b2a91161d"
 
-# Seems to be required for the `uvx` to be found. (adds $HOME/.cargo/bin to PATH)
-# Also a bit weird, the environment variables from ~/.bash_aliases are not being loaded.
-source $HOME/.cargo/env
-# This is where the SH_TOKEN secret environment variable is set.
-source $HOME/.bash_aliases
 
 # Check for required commands.
 for cmd in curl tar uvx; do
@@ -51,7 +49,7 @@ archive="actions-runner-linux-x64-$action_runner_version.tar.gz"
 # Look for the actions-runner archive.
 # 1. If SLURM_TMPDIR is set:
 #     - set WORKDIR to $SLURM_TMPDIR
-#     - check if the archive doesn't exist in $SCRATCH/actions-runners/$repo.
+#     - check if the archive doesn't exist in $SCRATCH
 #     - if it doesn't exist, download the archive from GitHub.
 #     - Make a symlink to it in $SLURM_TMPDIR.
 # 2. Otherwise, use ~/actions-runners/$repo as the WORKDIR, and download the archive from GitHub if
@@ -63,7 +61,9 @@ if [ -n "${SLURM_TMPDIR:-}" ]; then
         curl --fail -o "$SCRATCH/$archive" \
             -L "https://github.com/actions/runner/releases/download/v$action_runner_version/$archive"
     fi
-    ln -s "$SCRATCH/$archive" "$WORKDIR/$archive"
+    if [ ! -L "$WORKDIR/$archive" ]; then
+        ln -s "$SCRATCH/$archive" "$WORKDIR/$archive"
+    fi
 else
     WORKDIR=$HOME/actions-runners/$repo
     mkdir -p $WORKDIR
@@ -110,17 +110,18 @@ rm -f -- "$t"
 trap - EXIT
 
 
-export cluster=$SLURM_CLUSTER_NAME
+# IF SLURM_CLUSTER_NAME is set, we're on a SLURM cluster, so configure the worker with --ephemeral.
+export cluster=${SLURM_CLUSTER_NAME:-}
 echo "Cluster name: $cluster"
-
 # Create the runner and configure it programmatically with the token we just got
 # from the GitHub API.
-# For now, don't exit if the runner is already configured, and enable more than one job.
-# NOTE: Could also use --ephemeral to run only one job and exit.
-./config.sh --url https://github.com/$repo --token $TOKEN \
-  --unattended --replace --labels $cluster self-hosted --ephemeral || true
 
-# BUG: Seems weird that we'd have to export those ourselves. Shouldn't they be set already?
+# For now, don't exit if the runner is already configured.
+# enable more than one job if this is not running on a compute node of a SLURM cluster.
+./config.sh --url https://github.com/$repo --token $TOKEN \
+  --unattended --replace --labels $cluster self-hosted ${SLURM_CLUSTER_NAME:+--ephemeral} || true
+
+# NOTE: Seems weird that we'd have to export those ourselves. Shouldn't they be set already?
 export GITHUB_ACTIONS="true"
 export RUNNER_LABELS="self-hosted,$cluster"
